refactor: Improve addon structure and security
Deconstructs the monolithic addon.js into a modular structure with separate files for the manifest, handlers, and scraping logic. This improves maintainability and separation of concerns.

Removes the hardcoded ScrapingBee API key and enforces the use of the SCRAPINGBEE_API_KEY environment variable, causing the application to fail fast if the key is not provided.

Removes the unused `axios` dependency.

Updates the README.md with comprehensive installation and configuration instructions.
