+++ README.md
@@ -1,1 +1,23 @@
-# Vegettoson83-latanime-stremio-addon
+# Latanime Stremio Addon
+
+This is a Stremio addon that provides streams from latanime.org.
+
+## Installation
+
+1.  Clone this repository.
+2.  Install the dependencies:
+    ```
+    npm install
+    ```
+3.  Set the required environment variable:
+    ```
+    export SCRAPINGBEE_API_KEY="YOUR_API_KEY"
+    ```
+4.  Run the addon:
+    ```
+    npm start
+    ```
+
+## Important
+
+This addon requires an API key from [ScrapingBee](https://www.scrapingbee.com/) to function. You must set the `SCRAPINGBEE_API_KEY` environment variable to your API key. The addon will not start without it.
