+++ package.json
@@ -7,7 +7,6 @@
     "start": "node server.js"
   },
   "dependencies": {
-    "axios": "^1.7.2",
     "cheerio": "^1.0.0-rc.12",
     "node-cache": "^5.1.2",
     "scrapingbee": "^1.7.6",
